backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
coverage==7.4.4
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-mail==1.4.1
freezegun==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.26.0
rsa==4.9
ruff==0.3.4
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.28.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
